// packages/leeloo/src/webapp/account/types.ts
export enum ActivePanel {
    Root = 'root',
    AccountDetails = 'account-details',
    DeviceManagement = 'device-management',
    SecuritySettings = 'security-settings',
    ImportExportData = 'import-export-data',
    AccessAllFeatures = 'access-all-features',
    AccountRecovery = 'account-recovery'
}